i9300: libsensors: Fix compass orientation (for real)
Actually this fixes several more problems, but the compass orientation
is the most noticeable symptom of this:

Android expects sensor event timestamps in nanoseconds. Linux timestamps
on the other hand are saved in the timeval struct in seconds and micro-
seconds. Therefore these need to be converted, which is easy enough and
is implemented in input.c of libsensors. Unfortunately, the current
implementation uses type long int for the return value and all further
calculations. Since the nanosecond values easily exceed 2^32 the buffers
overflow and wrap, which results in the reported timestamps moving between
-2^16 and 2^16. The consequences of this are noticeable especially with
fused sensors integrating the gyroscope readings, which use the difference
of two timestamps to calculate the change in orientation by multiplying it
by the rotation rate. When the buffer wraps around this results in a huge
time difference and thus leads to a momentary orientation change every
~4.3 seconds.

This commit fixes this behavior by using long long int.

+++ libsensors/akm8975.c
@@ -168,7 +168,7 @@ void *akm8975_thread(void *thread_data)
 	char i2c_data[SENSOR_DATA_SIZE] = { 0 };
 	short magnetic_data[3];
 	short mode;
-	long int before, after;
+	long long int before, after;
 	int diff;
 	int device_fd;
 	int uinput_fd;
